Major Refactor On Tests
Tests now pass in the account related to the specific role.

These roles are added and tested at the beginning of the contract.

The tests are also more relevant now to the project, rather than testing the same criteria, they look for the changes to the state and specific values updated/changed